usb: Remove unnecessary input current limitation in fast charge.
The fast charge patch changes SDP input current to 1500mA. Then it was limited to 1200mA here.
Additionally, this limits CDP and DCP (AC charger) input current to 1200mA in fast charge mode, where with it disabled it defaults to 1500mA with option to go up to 2000mA for high power AC adapters.